+++ Intent.Module.TypeScript.Domain/release-notes.md
@@ -1,3 +1,4 @@
 ### Version 3.3.5
 
 - Fixed: Superfluous leading whitespace would sometimes be added to decorators on entities.
+- `Is Abstract` is now respected on domain entities.